pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "declargs"
authors = [{name = "Lucas Viana", email = "lucas.nviana@gmail.com"}]
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]
dependencies = [
    "pyyaml >=2.6",
]

[project.optional-dependencies]
dev = [
    "tox",
]

[project.urls]
Home = "https://www.lucas-viana.com"
